dbDatabase::attach() must have a corresponding dbDatabase::detach() ...
...in order to clear the transaction locks.  Otherwise, the transaction
continues to exist in the database, and others will block waiting 
for it to complete.

Fixes XECS-619 - IMDB semaphores get out of whack after soak tests.

The reason it took a long soak test for this to show up, is because
the code that called SubscriptionDB::subscriptionExists(), would
usually call SubscriptionDB::insertRow(), which called detach().  When
a subscription finally expired, and SubscriptionDB::subscriptionExists()
returned with no subscription, THEN the code wouldn't call
SubscriptionDB::insertRow(), and the transaction was left open.  In the
soak tests, the subscription duration was 15 minutes, so it took that
long for the system to follow the error path.

If the subscriptions were refreshed before they expired, this bug would 
never have been detected.

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11453 ab1d8caa-1f67-47f1-9e81-24633a41865c