Call exit() if the IMDB database fails to open.
If an application cannot open the IMDB, it cannot proceed.  Log
a critical error and call exit().

Only added to SIPDBManager::getDatabase(), which is what applications
use.  The internally used SIPDBManager::openDatabase() is not changed,
as it may be important for internal use to handle this error differently.

Fixes XECS-643 - sipX applications fail to initialize IMDB and yet stay up -
all registrations fail

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11490 ab1d8caa-1f67-47f1-9e81-24633a41865c