sipXpublisher: Remove cruft from POST body befor sending to phones.
The way sipXivr uses a POST request to sipstatus, instead of a GET
with a body as sipXvxml used to do, caused some extra lines to
appear in the application/simple-message-summary NOTIFY sent to the 
phones.

Most phones appear to ignore them.  The SMC3456 did not, and responded
with a 400 Bad Request.

Changed MwiPlugin.cpp to remove everything before the "messages-waiting:" 
part of the body so the extra stuff isn't sent to the phone.

Fixes XX-6887 - SMC3456:Delay (Approx 3min) in MWI indication upon depositing/retrieving voice mail for SMC user.

The reason for the delay was that while the initial NOTIFY message was 
rejected (the one pushed by sipXivr), the SUBSCRIBE was refreshed and when 
that occured the correct body was sent and the MWI was accepted.

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@17088 ab1d8caa-1f67-47f1-9e81-24633a41865c