PT45302941 - High memory usage for sipxregristrar
The registrar was leaking memory because of the way mongo scoped connections were used.
After the C++ driver was updated to 2.2.x a scoped connection is owned by the caller
and has to be deleted. All scoped connections were wrapped in a boost::scoped_ptr to
guarantee that they are deleted when they go out of scope.