UC-3263 Call park retrieve through BLF does not work anymore
Redone the SipRedirectorPickUp code related to retrieving calls from park orbits to use both
call pickup (default *78)  and call retrieve (default *4) as call retrieval codes.
If the remainder of the userid after the call pickup/call retrieval code is a park orbit the
SipRedirectorPickUp stops processing and lets the SipRedirectorAliasDB to retrieve the park orbit
FS contact from mongo imdb. In case call pickup code is detected but the userid is not a park
orbit nothing happens, and the redirector continues with the regular call pickup logic.
In case call retrieval code is detected but the remainder of the userid is not a park orbit
redirector will return a 404.